requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-ckeditor==6.7.0
django-redis==5.4.0
googletrans==3.1.0a0
redis==5.0.1